Add files via upload
The PS1 file is not necessary to use the addon.  I am not as good with lua code as powershell, so I wrote the script in powershell and used AI to help convert it to lua.  The ps1 file is the original powershell script I wrote.  I created the logic including the $moonphases variable.